fix: change task_id and user_id types from INTEGER to BIGINT for consistency

+++ database/models/task.js
@@ -3,12 +3,12 @@ module.exports = (sequelize, DataTypes) => {
     'Task',
     {
       task_id: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT,
         autoIncrement: true,
         primaryKey: true,
       },
       user_id: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT,
         allowNull: false,
       },
       estimate: {
